refactor dendrocore to separate package
